Move waitingGame to its own file
Until we decide on an organisation for our frontend code, I think it makes more sense to be consistent with the rest of the code, so I’m putting this functionality in its own module.
